django-ninja-server/requirements.txt
django==4.2.7
daphne==4.0.0
django-ninja==1.0.1
mysqlclient==2.2.0